tests/submission/W3C/domhtml/HTMLBodyElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-body-text" rel="help">
<link rel="author" title="NIST">
</head>
<body text="#000000"> 
<p>Hello, World</p>


<script>
var domtest = async_test("The text IDL attribute of the body element must reflect the element's text content attribute.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtext;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("body");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtext = testNode.text;

      assert_equals(vtext,"#000000");
       
    domtest.done();
}));

</script>
<p>The text IDL attribute of the body element must reflect the element's text content attribute.</p>
<div id="log"></div>
</body></html>
